+++ FidoDidoGame/Modules/Rank/Request/UpdateRank.cs
@@ -4,13 +4,13 @@
     {
         public string? UserName { get; set; }
 
-        public int UserId { get; set; }
+        public long UserId { get; set; }
 
         public int Point { get; set; }
 
         public DateTime Date { get; set; }
 
-        public UpdateRank(string? userName, int userId, int point, DateTime date)
+        public UpdateRank(string? userName, long userId, int point, DateTime date)
         {
             UserName = userName;
             UserId = userId;
